# Use slicing by 8 more aggressively.