--- a/crc32_amd64.go
+++ b/crc32_amd64.go
@@ -15,10 +15,12 @@ func haveCLMUL() bool
 
 // castagnoliSSE42 is defined in crc_amd64.s and uses the SSE4.2 CRC32
 // instruction.
+// go:noescape
 func castagnoliSSE42(crc uint32, p []byte) uint32
 
 // ieeeCLMUL is defined in crc_amd64.s and uses the PCLMULQDQ
 // instruction as well as SSE 4.1.
+// go:noescape
 func ieeeCLMUL(crc uint32, p []byte) uint32
 
 var sse42 = haveSSE42()
@@ -27,6 +29,10 @@ var useFastIEEE = haveCLMUL() && haveSSE41()
 func updateCastagnoli(crc uint32, p []byte) uint32 {
 	if sse42 {
 		return castagnoliSSE42(crc, p)
+	}
+	// only use slicing-by-8 when input is >= 16 Bytes
+	if len(p) >= 16 {
+		return updateSlicingBy8(crc, castagnoliTable8, p)
 	}
 	return update(crc, castagnoliTable, p)
 }
@@ -42,8 +48,8 @@ func updateIEEE(crc uint32, p []byte) uint32 {
 		return crc
 	}
 
-	// only use slicing-by-8 when input is >= 4KB
-	if len(p) >= 4096 {
+	// only use slicing-by-8 when input is >= 16 Bytes
+	if len(p) >= 16 {
 		iEEETable8Once.Do(func() {
 			iEEETable8 = makeTable8(IEEE)
 		})