Use slicing by 8 for smaller IEEE sizes, and use it for Castagnoli tables as well, if there is no SSE 4.2 available.

Slicing-by-8 seems to be faster from around 16-20 byte size slices and upwards. 40 bytes is approximately 1.5x the speed.

Non SSE 4.2 performance:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkCastagnoli40B-4      161           105           -34.78%
BenchmarkCastagnoli1KB-4      3603          1505          -58.23%
BenchmarkCastagnoli8KB-4      28741         12103         -57.89%
BenchmarkCastagnoli32KB-4     114681        47015         -59.00%

benchmark                     old MB/s     new MB/s     speedup
BenchmarkCastagnoli40B-4      247.50       378.79       1.53x
BenchmarkCastagnoli1KB-4      284.19       680.36       2.39x
BenchmarkCastagnoli8KB-4      285.02       676.82       2.37x
BenchmarkCastagnoli32KB-4     285.73       696.96       2.44x
```

IEEE 40 byte/update speed:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkCrc40B-4     161           110           -31.68%

benchmark             old MB/s     new MB/s     speedup
BenchmarkCrc40B-4     247.95       361.54       1.46x
```